interface ts renamed
