+++ src/app/interfaces/accordion.interface.ts
